+++ src/main/java/com/firegodjr/ancientlanguage/api/magic/package-info.java
@@ -1,0 +1,4 @@
+//@API(owner = com.firegodjr.ancientlanguage.Main.MODID, provides = "MagicHandling", apiVersion = "0.1")
+package com.firegodjr.ancientlanguage.api.magic;
+
+//import net.minecraftforge.fml.common.API;
