+++ src/main/java/com/firegodjr/ancientlanguage/api/script/events/package-info.java
@@ -1,0 +1,4 @@
+@API(owner = com.firegodjr.ancientlanguage.Main.MODID, provides = "ScriptEvents", apiVersion = "0.1")
+package com.firegodjr.ancientlanguage.api.script.events;
+
+import net.minecraftforge.fml.common.API;
