+++ src/main/java/com/firegodjr/ancientlanguage/magic/words/WordAction.java
@@ -226,5 +226,11 @@ public class WordAction {
 				}
 			}
 		}
+
+		@Override
+		public Entity onWardPlace(MagicData data, Map<String, String> modData,
+				Vec3 position, String scriptToInsert) {
+			return null;
+		}
 	}
 }
