+++ src/main/java/test/firegodjr/ancientlanguage/TypeActiveEventTest.java
@@ -4,7 +4,7 @@ import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 
 import com.firegodjr.ancientlanguage.api.script.IScriptObject;
-import com.firegodjr.ancientlanguage.event.ScriptEvent;
+import com.firegodjr.ancientlanguage.api.script.events.ScriptEvent;
 
 public class TypeActiveEventTest {
